Deleted some code that is not relevant to this fork of the original repository.
Also changed the name of the malloc_count header file. I am going to change
the way that we intercept calls to malloc to use a preprocessor macro
called malloc() instead of overriding the function. That way, the macro
can add in info like the function, file and line where malloc or free is
being called.