Fixed errors from test run. Turns out that the __LINE__ macro returns the line number as an int, not a char *, so changed the printf format string to stop it from dumping core.
Also, retrieving the address from the pointer does not seem to work. I
know that free() does it, but free() was written by people who grok malloc
and it's data structures. Instead, my malloc wrapper allocates a bit of
extra padding and stores the length in that padding. This will work just
fine as long as programs don't go crazy with pointer arithmetic.